Revised the bare repository approach
Instead of merely using the `--work-tree` feature of Git, for the
bare repository approach, also use the `--git-dir`.
This means the home directory is no longer considered a Git repository,
which provides the following benefits,

  - No Git confusion in terms of global configuration (~/.gitconfig) vs. local configuration (~/.git/config).
  - No risk of accidental Git operations on dotfiles, that were intended
for other repositories in the home directory.

The only downside is that the normal Git commands are no longer
valid, unless the `--gitdir` and `--work-tree` options are provided.
The wrapper script `bin/dotfiles` is provided for this purpose.

The installation instructions have been revised to describe this,
and also cleaned up a bit.